Changed Database Structure

+++ HotelManagementSystem/src/main/java/com/mikeshaggy/hms/amenity/repository/AmenityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 //import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 //@RepositoryRestResource(path = "amenities")
-public interface AmenityRepository extends JpaRepository<Amenity, Integer> {
+public interface AmenityRepository extends JpaRepository<Amenity, Long> {
 }
